added autharization

+++ backend/routes/post.js
@@ -4,6 +4,7 @@ const router = express.Router();
 const Post = require('../models/post')
 //const router = express.Router()
 const mongoose = require('mongoose');
+const checkAuth = require('../middleware/check_auth')
 
 const MIME_TYPE ={
   'image/png': 'png',
@@ -29,27 +30,30 @@ filename: (req,file, cb) =>
 }
   })
 
-router.post('', multer({storage: storage}).single("image"), (req,res,next)=>{
+router.post('', checkAuth,  multer({storage: storage}).single("image"), (req,res,next)=>{
   const url = req.protocol+'://'+req.get("host");
+ // console.log(url);
   const post = new Post({
     _id: req.params.id,
     title: req.body.title,
     content: req.body.content,
     date:new Date(req.body.date),
-    imagePath: url+ "/images/" + req.file.filename
+    imagePath: url+ "/images/" + req.file.filename,
+    creator: req.userData.userId
   })
   //const post= req.body;
  post.save().then(createdpost =>
     {
+      console.log("saved")
       res.status(201).json({
         message: 'Post added succesfully',
         id:createdpost._id,
-        post:createdpost.imagePath
+        p:createdpost.imagePath
       });
     });
  })
 
-router.put('/:id', multer({storage: storage}).single("image") , (req,res,next)=>
+router.put('/:id', checkAuth, multer({storage: storage}).single("image") , (req,res,next)=>
 {
 let imagePath= req.body.imagePath;
 if(req.file)
@@ -60,49 +64,47 @@ if(req.file)
   Post.findById(req.params.id).then(p=>{
     if(p)
     {
-      // res.status(200).json(p);
 
+      if(p.creator == req.userData.userId){
       p.title = req.body.title
       p.content= req.body.content
       p.date = new Date(req.body.date)
       p.imagePath = imagePath
+
       console.log(p);
       p.save().then(result=>{
         console.log(result);
         res.status(200).json({message:"update Succeful!", image:imagePath});
       }).catch((p)=>{
           console.log("error", p);})
+      }
+      else
+    {
+      res.status(401).json({message:"unautherize"})
+    }
     }
     else
     {
-      res.json(400).json({message:"unable to fetch"})
+      res.status(400).json({message:"unable to fetch"})
     }
   })
-  //const post = new Post({
-  //   // _id: req.params.id,
-  //   title: req.body.title,
-  //   content: req.body.content
-  // });
-  // var updatedPost = {};
-  // console.log(req.params.id);
-  // // updatedPost = Object.assign(updatedPost, req.body);
-  // // Post.updateOne({_id: req.params.id},{$set: {updatedPost}}).then(result =>{
-  //   Post.updateOne({_id: req.params.id},{$set:{post}}).then(result =>{
-  //   console.log("hi");
-  //   console.log(result);
-  //   res.status(200).json({message:"update Succeful!"});
-  // }).catch((p)=>{
-  //   console.log("error", p);
-  // })
+
 })
 
-router.delete('/:id', (req,res,next)=>
+router.delete('/:id', checkAuth, (req,res,next)=>
 {
 
-  Post.deleteOne({_id: req.params.id}).then(
+  Post.deleteOne({_id: req.params.id, creator:req.userData.userId}).then(
     result => {
       console.log(result);
+      if(result.n>0){
+      console.log(result);
       res.status(200).json({message: 'deleted'})
+    }
+    else
+    {
+      res.status(401).json({message:"unautherized"})
+    }
     }
   )
  // console.log("hi");
@@ -121,7 +123,7 @@ router.get('/:id', (req,res,next)=>
     }
   })
 })
-router.use('',(req, res, next)=>
+router.get('', (req, res, next)=>
 {
   const pageSize = +req.query.pageSize;
   const currpage = +req.query.page;
@@ -130,6 +132,8 @@ router.use('',(req, res, next)=>
   if(pageSize && currpage){
     postQuery.skip(pageSize * (currpage - 1)).limit(pageSize);
   }
+  if(!postQuery)
+    res.status(200).json({message:"no post"})
   postQuery.then(doc =>
     {
       fetchedPost =doc;
